+++ packages/stream_video_flutter/lib/src/call_screen/call_content/picture_in_picture/stream_picture_in_picture_ui_kit_view.dart
@@ -93,7 +93,7 @@ class _StreamPictureInPictureUiKitViewState
       compare:
           widget.pictureInPictureConfiguration?.sort ??
           widget.participantSort ??
-          CallParticipantSortingPresets.speaker,
+          CallParticipantSortingPresets.pictureInPicture,
     );
 
     if (sorted.isNotEmpty) {