+++ packages/stream_video_noise_cancellation/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.2.2
+* Sync version with `stream_video_flutter` 1.2.2
+
+## 1.2.1
+* Sync version with `stream_video_flutter` 1.2.1
+
 ## 1.2.0
 * Sync version with `stream_video_flutter` 1.2.0